Set the correct template to the upsell filter on category level
For the recommendations it's possible to set a template on global
level (Magento configuration), product level (attribute) and category
level (attribute). For the category an incorrect source model was used.
It uses `emico_tweakwise/system_config_source_recommendationFeatured`,
where it should use `emico_tweakwise/system_config_source_recommendationProduct`.
With a new install script this bug can be fixed, so it's possible
to select the correct upsell template on category level as well.